[routes] rename route and component
