+++ src/pages/DevPage/index.ts
@@ -1,0 +1,3 @@
+import { DevPage } from './DevPage'
+
+export default DevPage();
