--- src/pages/SettingsPage/data.ts
@@ -1,1 +1,0 @@
-export const settingsData = {}
