+++ src/types/Paths.ts
@@ -2,7 +2,7 @@ export enum Paths {
   home = '/',
   signIn = '/signin',
   signUp = '/signup',
-  settings = '/settings',
+  dev = '/dev',
   error404 = "/error404",
   error500 = "/error500",
   pageNotFound = "/page-not-found",
